+++ packages/tiptap-editor/web/src/commands/deleteNode.ts
@@ -12,11 +12,12 @@ declare module '@tiptap/core' {
   }
 }
 
-export const deleteNode: RawCommands['deleteNode'] = id => ({ view, tr }) => {
+export const deleteNode: RawCommands['deleteNode'] = id => ({ view, tr, dispatch }) => {
   const nodeWithPos = findChildren(view.state.doc, node => node.attrs.id === id);
   const { pos, node } = nodeWithPos?.[0] || {};
-  if (pos !== undefined) {
-    view.dispatch(tr.delete(pos, pos + node.nodeSize));
+  if (pos !== undefined && dispatch) {
+    tr.delete(pos, pos + node.nodeSize);
+
     return true;
   } else {
     return false;