+++ packages/viewer/web/src/List.jsx
@@ -53,11 +53,15 @@ const List = ({
           alignment
         );
         const hasJustifyText = alignment === 'justify' && hasText(children);
+        const shouldFixHeight = context.experiments?.fixListLineHeight?.enabled;
+        const blockStyle = blockDataToStyle(blockProps.data[childIndex]);
+        const style = shouldFixHeight ? { lineHeight: blockStyle.lineHeight } : undefined;
         const elementProps = key => ({
           className: classNames(mergedStyles.elementSpacing, textClassName, {
             [styles.hasJustifyText]: hasJustifyText,
             [styles.contentCenterAlignment]: alignment === 'center',
           }),
+          style,
           key,
         });
         React.Children.forEach(children, (child, i) => {
@@ -100,7 +104,8 @@ const List = ({
               styles[alignment],
               getBlockStyleClasses(mergedStyles, listItemDirection, alignment, className, true),
               isPaywallSeo(context.seoMode) &&
-                getPaywallSeoClass(context.seoMode.paywall, blockIndex)
+                getPaywallSeoClass(context.seoMode.paywall, blockIndex),
+              shouldFixHeight && styles.lineHeightFix
             )}
             key={blockProps.keys[childIndex]}
             style={blockDataToStyle(blockProps.data[childIndex])}
@@ -123,6 +128,7 @@ List.propTypes = {
   textDirection: PropTypes.oneOf(['rtl', 'ltr']),
   context: PropTypes.shape({
     theme: PropTypes.object.isRequired,
+    experiments: PropTypes.object,
     anchorTarget: PropTypes.string.isRequired,
     relValue: PropTypes.string.isRequired,
     config: PropTypes.object.isRequired,